requirements.txt
biopython
receptor-utils
